// C1220G2-Sprint2-Angular-Source/angular-sprint2/src/app/models/teacher.ts
export interface Teacher {
  code: string;
  name: string;
  dateOfBirth: string;
  gender: string;
  phone: string;
  email: string;
  address: string;
  image: string;
  twitter: string;
  facebook: string;
  facultyId: any;
  educationId: any;
}

export class Teacher {
  code: string;
  name: string;
  dateOfBirth: string;
  gender: string;
  phone: string;
  email: string;
  address: string;
  image: string;
  twitter: string;
  facebook: string;
  facultyId: any;
  educationId: any;
}